--- a/.github/workflows/publish.yml
+++ b/.github/workflows/publish.yml
@@ -5,8 +5,46 @@ on:
       - master
 
 jobs:
+  sanity-check:
+    name: Pre-publish checks
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Use Node.js 12.x
+        uses: actions/setup-node@v2
+        with:
+          node-version: 12.x
+      - name: Get Date
+        id: get-date
+        shell: bash
+        run: echo "::set-output name=date::$(/bin/date -u "+%Y%m%d")"
+      - name: Cache dependencies
+        uses: actions/cache@v2
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-
+      - name: Cache Jest
+        uses: actions/cache@v2
+        with:
+          path: .cache/jest
+          key: ${{ runner.os }}-jest-${{ hashFiles('package-lock.json') }}-${{ steps.get-date.outputs.date }}
+          restore-keys: |
+            ${{ runner.os }}-jest-${{ hashFiles('package-lock.json') }}-
+            ${{ runner.os }}-jest-
+      - name: Install dependencies
+        run: npm ci
+      - name: Build NodeJS package
+        run: npm run build
+      - name: Run linters
+        run: npm run lint
+      - name: Run tests
+        run: npm run test
   npm:
     name: NPM Package
+    needs: sanity-check
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -24,16 +62,13 @@ jobs:
             ${{ runner.os }}-node-
       - name: Install dependencies
         run: npm ci
-      - name: Sanity check
-        run: |
-          npm run lint
-          npm run test
       - name: Deploy to NPM
         uses: JS-DevTools/npm-publish@v1
         with:
           token: ${{ secrets.NPM_TOKEN }}
   github:
     name: GitHub release
+    needs: sanity-check
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -51,10 +86,6 @@ jobs:
             ${{ runner.os }}-node-
       - name: Install dependencies
         run: npm ci
-      - name: Sanity check
-        run: |
-          npm run lint
-          npm run test
       - name: Get commit message (for release title and body)
         id: commit
         uses: kceb/git-message-action@v1