--- a/.github/workflows/verify.yml
+++ b/.github/workflows/verify.yml
@@ -44,6 +44,10 @@ jobs:
         uses: actions/setup-node@v2
         with:
           node-version: 12.x
+      - uses: actions/cache@v2
+        with:
+          path: .jestcache
+          key: jest-${{ hashFiles('package-lock.json') }}
       - name: Install dependencies
         run: npm ci
       - name: Run tests