--- a/.github/workflows/verify.yml
+++ b/.github/workflows/verify.yml
@@ -60,9 +60,8 @@ jobs:
           node-version: 12.x
       - name: Get Date
         id: get-date
-        run: |
-          echo "::set-output name=date::$(/bin/date -u "+%Y%m%d")"
         shell: bash
+        run: echo "::set-output name=date::$(/bin/date -u "+%Y%m%d")"
       - name: Cache dependencies
         uses: actions/cache@v2
         with:
@@ -70,11 +69,14 @@ jobs:
           key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
           restore-keys: |
             ${{ runner.os }}-node-
-      - name: Cache Jest's cache
+      - name: Cache Jest
         uses: actions/cache@v2
         with:
-          path: .jestcache
-          key: jest-${{ steps.get-date.outputs.date }}-${{ hashFiles('package-lock.json') }}
+          path: .cache/jest
+          key: ${{ runner.os }}-jest-${{ hashFiles('package-lock.json') }}-${{ steps.get-date.outputs.date }}
+          restore-keys: |
+            ${{ runner.os }}-jest-${{ hashFiles('package-lock.json') }}-
+            ${{ runner.os }}-jest-
       - name: Install dependencies
         run: npm ci
       - name: Run tests