--- a/.github/workflows/verify.yml
+++ b/.github/workflows/verify.yml
@@ -43,11 +43,7 @@ jobs:
       - name: Install dependencies
         run: npm ci
       - name: Run linter
-        run: |
-          npm run jsonlint
-          npm run svglint
-          npm run wslint
-          npm run our-lint
+        run: npm run lint
   test:
     name: Test package
     runs-on: ubuntu-latest