--- a/.gitignore
+++ b/.gitignore
@@ -19,6 +19,9 @@ lerna-debug.log*
 # Dependency directories
 node_modules/
 
+# Jest
+.jestcache
+
 
 ### macOS ###
 # General