… icons/42.svg