@raeffs 2021-06-12
@ericcornelissen just saw this PR and had a look at the workflows as I was wondering what you are doing during the CI runs. Have you considered other optimizations? For example you could do a git diff and just run the test for icons that were affected by changes. 🙂