@ericcornelissen 2021-06-12
Nothing specific @raeffs, but suggestions are always welcome :slightly_smiling_face: 

How could we go about implementing that suggestion? If you have a robust and simple solution I definitely think we can do that :+1: