@ericcornelissen 2021-06-16
In the spirit of further optimizations, I included dependency caching in this Pull Request. Again, it doesn't seem to help us much (given how relatively few dependencies we have) but it's something, about 1 to 2 seconds per job.